00-PROGRAM/platform/llm/gateway/requirements.txt
# LLM Gateway Requirements
# Multi-provider LLM gateway with security and observability

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.3

# HTTP client for provider APIs
httpx>=0.25.0
aiohttp>=3.9.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# KMS integration
boto3>=1.28.0  # AWS KMS
azure-identity>=1.14.0
azure-keyvault-secrets>=4.7.0
google-cloud-kms>=2.19.0

# Observability
structlog>=23.2.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Rate limiting
slowapi>=0.1.9

# Validation
jsonschema>=4.19.0
